Add constructors to models

+++ src/main/java/com/example/aucaregistration/repository/ClientRepository.java
@@ -3,4 +3,4 @@ package com.example.aucaregistration.repository;
 import com.example.aucaregistration.domain.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientRepository extends JpaRepository<Client, Long> { }
+public interface ClientRepository extends JpaRepository<Client, Integer> { }
